requirements.txt
pygame-ce>=2.5.6
Pillow>=12.1.1
